+++ equalKeys_test.go
@@ -1,6 +1,11 @@
-package fosite
+package fosite_test
 
-import "testing"
+import (
+	"testing"
+	"github.com/oleiade/reflections"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
 
 func TestAssertObjectsAreEqualByKeys(t *testing.T) {
 	type foo struct {
@@ -14,4 +19,47 @@ func TestAssertObjectsAreEqualByKeys(t *testing.T) {
 	AssertObjectKeysEqual(t, a, a, "Name", "Body")
 	AssertObjectKeysNotEqual(t, a, b, "Name")
 	AssertObjectKeysNotEqual(t, a, c, "Name", "Body")
+}
+
+func AssertObjectKeysEqual(t *testing.T, a, b interface{}, keys ...string) {
+	assert.True(t, len(keys) > 0, "No keys provided.")
+	for _, k := range keys {
+		c, err := reflections.GetField(a, k)
+		assert.Nil(t, err)
+		d, err := reflections.GetField(b, k)
+		assert.Nil(t, err)
+		assert.Equal(t, c, d, "%s", k)
+	}
+}
+
+func AssertObjectKeysNotEqual(t *testing.T, a, b interface{}, keys ...string) {
+	assert.True(t, len(keys) > 0, "No keys provided.")
+	for _, k := range keys {
+		c, err := reflections.GetField(a, k)
+		assert.Nil(t, err)
+		d, err := reflections.GetField(b, k)
+		assert.Nil(t, err)
+		assert.NotEqual(t, c, d, "%s", k)
+	}
+}
+
+func RequireObjectKeysEqual(t *testing.T, a, b interface{}, keys ...string) {
+	assert.True(t, len(keys) > 0, "No keys provided.")
+	for _, k := range keys {
+		c, err := reflections.GetField(a, k)
+		assert.Nil(t, err)
+		d, err := reflections.GetField(b, k)
+		assert.Nil(t, err)
+		require.Equal(t, c, d, "%s", k)
+	}
+}
+func RequireObjectKeysNotEqual(t *testing.T, a, b interface{}, keys ...string) {
+	assert.True(t, len(keys) > 0, "No keys provided.")
+	for _, k := range keys {
+		c, err := reflections.GetField(a, k)
+		assert.Nil(t, err)
+		d, err := reflections.GetField(b, k)
+		assert.Nil(t, err)
+		require.NotEqual(t, c, d, "%s", k)
+	}
 }
