+++ glide.yaml
@@ -13,11 +13,6 @@ import:
   version: ~1.0.0
 - package: github.com/pkg/errors
   version: ~0.8.0
-- package: github.com/stretchr/testify
-  version: ~1.1.4
-  subpackages:
-  - assert
-  - require
 - package: golang.org/x/crypto
   subpackages:
   - bcrypt
@@ -29,3 +24,8 @@ testImport:
 - package: golang.org/x/oauth2
   subpackages:
   - clientcredentials
+- package: github.com/stretchr/testify
+  version: ~1.1.4
+  subpackages:
+  - assert
+  - require
